% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{arrayPlot}
\alias{arrayPlot}
\title{Array pie plot}
\usage{
arrayPlot(scores, coordinates, image = NULL, spotScale = 1,
  spotOpacity = 1)
}
\arguments{
\item{scores}{matrix of dimension (n, k), where n is the number of data
points and k is the number of categories that are displayed in the pie
charts.}

\item{coordinates}{data.frame of dimension (n, 2) encoding the plotting
positions of the n data points. \code{coordinates} should contain the columns 'x'
and 'y' and its \code{rownames} should correspond to those in \code{scores}.}

\item{image}{a grob to display as a background image to the plot.}

\item{spotScale}{size of the pie charts.}

\item{spotOpacity}{opacity of the pie charts.}
}
\value{
ggplot object of the array pie plot.
}
\description{
Creates the pie plot.
}
\keyword{internal}
