% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{clusterTree}
\alias{clusterTree}
\title{Cluster tree}
\usage{
clusterTree(assignments, transitionProportions = "To",
  transitionLabels = FALSE, transitionThreshold = 0)
}
\arguments{
\item{transitionProportions}{how to compute and display edge labels. Valid values
are:
\itemize{
\item \code{"To"}: Proportions are computed based on the number of features in the
higher resolution cluster.
\item \code{"From"}: Proportions are computed based on the number of features in the
lower resolution cluster.
}}

\item{transitionLabels}{Show transition proportion labels.}

\item{graph}{graph of the clustering tree, as received from \code{CPieClust}.}
}
\value{
ggplot object of the clustering tree.
}
\description{
Creates the clustering tree.
}
\keyword{internal}
