% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.arrayPlot}
\alias{.arrayPlot}
\title{Array pie plot}
\usage{
.arrayPlot(scores, coordinates, image = NULL, spotScale = 1,
  spotOpacity = 1)
}
\arguments{
\item{scores}{(n, K) scoring matrix}

\item{coordinates}{\code{data.frame} with \code{rownames} matching those in \code{scores}
and columns \code{x} and \code{y} specifying the plotting position of each observation}

\item{image}{a \code{grob} to use as background to the plots}

\item{spotScale}{pie chart size}

\item{spotOpacity}{pie chart opacity}
}
\value{
\code{ggplot} object of the pie plot
}
\description{
Array pie plot
}
\keyword{internal}
