% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.computeClusterColors}
\alias{.computeClusterColors}
\title{Compute cluster colors}
\usage{
.computeClusterColors(clusterMeans)
}
\arguments{
\item{clusterMeans}{\code{data.frame} with feature means for each cluster.}
}
\value{
vector of cluster colors.
}
\description{
Computes colors so that dissimilar clusters are far away in color space.
}
\keyword{internal}
