% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.computeClusterDistances}
\alias{.computeClusterDistances}
\title{Compute cluster distances}
\usage{
.computeClusterDistances(assignments, counts)
}
\arguments{
\item{assignments}{list of assignment vectors.}

\item{counts}{count matrix.}
}
\value{
list of sample-centroid distances for each resolution.
}
\description{
Computes sample-centroid distances.
}
\keyword{internal}
