% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.makeApp}
\alias{.makeApp}
\title{SpatialCPie App}
\usage{
.makeApp(image, ...)
}
\arguments{
\item{image}{background image.}

\item{...}{arguments passed to \code{\link{.preprocessData}}.}
}
\value{
SpatialCPie \code{\link[shiny]{shinyApp}} object.
}
\description{
SpatialCPie App
}
\keyword{internal}
