% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.makeUI}
\alias{.makeUI}
\title{SpatialCPie UI}
\usage{
.makeUI(imageButton = FALSE)
}
\arguments{
\item{imageButton}{show image radio buttons}
}
\value{
SpatialCPie UI, to be passed to \code{shiny::shinyApp()}
}
\description{
SpatialCPie UI
}
\keyword{internal}
