% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.pairwiseDistance}
\alias{.pairwiseDistance}
\title{Pair-wise distances}
\usage{
.pairwiseDistance(a, b, d = function(x, ys) sqrt(colSums((t(ys) - x)^2)))
}
\arguments{
\item{a}{(n, d) matrix of points}

\item{b}{(m, d) matrix of points}

\item{d}{distance function (with signature [Num] -> [[Num]] -> [Num])}
}
\value{
(m, n) matrix of distances
}
\description{
Pair-wise distances
}
\keyword{internal}
