% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.SVGBarplot}
\alias{.SVGBarplot}
\title{SVG barplot}
\usage{
.SVGBarplot(xs)
}
\arguments{
\item{xs}{named vector with observations}
}
\value{
\code{\link{character}} SVG barplot
}
\description{
SVG barplot
}
\keyword{internal}
