% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.zscore}
\alias{.zscore}
\title{Z-score}
\usage{
.zscore(xs)
}
\arguments{
\item{xs}{vector of observations}
}
\value{
\code{xs}, z-normalized. if all elements of \code{xs} are equal, a vector of
zeros will be returned instead.
}
\description{
Z-score
}
\keyword{internal}
