% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{likeness}
\alias{likeness}
\title{Likeness score}
\usage{
likeness(d, c = 1)
}
\arguments{
\item{d}{distance matrix of dimension (n, k), where n is the number of data
points and k is the number of clusters.}

\item{c}{difference weighting constant. A higher value implies that higher
distances will be punished more. In the limit c -> ∞, the closest cluster
will be assigned a score of 1 and the other clusters a score of 0.}
}
\value{
(n, k) matrix of likeness scores.
}
\description{
Computes the normalized likeness score from a given distance matrix.
}
\keyword{internal}
