% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{pairwiseDistance}
\alias{pairwiseDistance}
\title{Pair-wise distances}
\usage{
pairwiseDistance(a, b, d = function(xs, y) sqrt(colSums((t(xs) - y)^2)))
}
\description{
Computes all pair-wise distances between two point clouds.
}
\keyword{internal}
