% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{runCPie}
\alias{runCPie}
\title{SpatialCPie gadget}
\usage{
runCPie(counts, clusterAssignments, img = NULL, view = "dialog",
  pixel.coords = NULL, transProp.threshold = 0.02)
}
\arguments{
\item{counts}{gene count matrix.}

\item{clusterAssignments}{list of cluster assignments for each resolution.}

\item{img}{image to be used as background to the plot (optional).\cr
Note: For the tissue image to be correctly aligned to the spatial areas, the
pixel.coords argument also needs to be provided.}

\item{view}{Shiny gadgets viewer options.
Available options: "dialog" (default), "browser", "pane".}

\item{pixel.coords}{\code{data.frame} with pixel coordinates (optional).
The rows should correspond to the columns (spatial areas) in the count file.}

\item{transProp.threshold}{threshold value for which transitions that are shown in the clustering tree.}
}
\value{
FIXME
}
\description{
Runs the SpatialCPie gadget
}
\examples{
options(device.ask.default = FALSE)

## set up coordinate system
coords <- as.matrix(expand.grid(1:10, 1:10))

## generate data set with three distinct genes
profiles <- diag(rep(1, 3)) + runif(9)
centers <- cbind(c(5, 2), c(2, 8), c(8, 2))
mixes <- array_branch(coords, 1) \%>\%
  map(~exp(-colSums((centers-.)^2)/50)) \%>\%
  map(~./sum(.)) \%>\%
  invoke(cbind, .)
means <- 100 * profiles \%*\% mixes
counts <- matrix(rpois(prod(dim(means)), means), nrow=nrow(profiles))
colnames(counts) <- array_branch(coords, 1) \%>\% map(~lift(paste)(., sep="x"))
rownames(counts) <- paste("gene", 1:nrow(counts))

## perform clustering
clusterAssignments <- 2:5 \%>\% map(~kmeans(t(counts), centers=.)$cluster)

## run SpatialCPie
runCPie(counts, clusterAssignments)
}
\keyword{arraypieplot}
\keyword{arrayplot}
\keyword{clustertree}
