% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.clusterTree}
\alias{.clusterTree}
\title{Cluster tree}
\usage{
.clusterTree(assignments, transitionProportions = "To",
  transitionLabels = FALSE, transitionThreshold = 0)
}
\arguments{
\item{assignments}{\code{\link[base]{data.frame}} with columns \code{"name"},
\code{"resolution"}, and \code{"cluster"}.}

\item{transitionProportions}{how to compute the transition proportions.
Possible values are:
\itemize{
\item \code{"From"}: based on the total number of assignments in the lower-resolution
cluster
\item \code{"To"}: based on the total number of assignments in the higher-resolution
cluster
}}

\item{transitionLabels}{\code{\link[base]{logical}} specifying whether to
show edge labels.}

\item{transitionThreshold}{hide edges with transition proportions below this
threshold.}
}
\value{
\code{\link[ggplot2]{ggplot}} object of the cluster tree.
}
\description{
Cluster tree
}
\keyword{internal}
