% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.computeClusterColors}
\alias{.computeClusterColors}
\title{Compute cluster colors}
\usage{
.computeClusterColors(clusterMeans)
}
\arguments{
\item{clusterMeans}{matrix of size \code{(n, K)} representing the \code{n} feature
means for each of the \code{K} clusters.}
}
\value{
vector of cluster colors.
}
\description{
Computes colors so that dissimilar clusters are far away in color space.
}
\keyword{internal}
