% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.computeClusterColors}
\alias{.computeClusterColors}
\title{Compute cluster colors}
\usage{
.computeClusterColors(assignments, counts)
}
\arguments{
\item{assignments}{list of assignment vectors.}

\item{counts}{count matrix.}
}
\value{
vector of cluster colors.
}
\description{
Computes colors so that dissimilar clusters are far away in color space.
}
\keyword{internal}
