% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.likeness}
\alias{.likeness}
\title{Likeness score}
\usage{
.likeness(d, c = 1)
}
\arguments{
\item{d}{distance vector.}

\item{c}{log multiplier.}
}
\value{
vector of scores.
}
\description{
Likeness score
}
\keyword{internal}
