% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.makeApp}
\alias{.makeApp}
\title{SpatialCPie App}
\usage{
.makeApp(counts, assignments, image = NULL, spotCoordinates = NULL)
}
\arguments{
\item{counts}{gene count matrix.}

\item{assignments}{cluster assignments.}

\item{image}{background image.}

\item{spotCoordinates}{spot pixel coordinates.}
}
\value{
SpatialCPie \code{shiny::shinyApp()} object
}
\description{
SpatialCPie App
}
\keyword{internal}
