% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.makeApp}
\alias{.makeApp}
\title{SpatialCPie App}
\usage{
.makeApp(counts, assignments, image = NULL, spotCoordinates = NULL)
}
\value{
SpatialCPie \code{shiny::shinyApp()} object
}
\description{
SpatialCPie App
}
