% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.makeServer}
\alias{.makeServer}
\title{SpatialCPie server}
\usage{
.makeServer(assignments, scores, colors, image, coordinates)
}
\arguments{
\item{colors}{vector of colors for each cluster label}

\item{image}{background image for the array plots, passed to
\code{\link[grid]{grid.raster}}}

\item{coordinates}{\code{\link[base]{data.frame}} with
\code{rownames} matching the \code{\link[base]{names}} in
\code{scores} and columns \code{x} and \code{y} specifying the plotting position of each
observation}

\item{distances}{list of spot-cluster distance matrices, one for each
resolution}
}
\value{
server function, to be passed to \code{\link[shiny]{shinyApp}}
}
\description{
SpatialCPie server
}
\keyword{internal}
