% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.makeServer}
\alias{.makeServer}
\title{SpatialCPie server}
\usage{
.makeServer(distances, colors, image, coordinates)
}
\arguments{
\item{distances}{list of spot-cluster distance matrices, one for each
resolution}

\item{colors}{vector of colors for each cluster label}

\item{image}{background image for the array plots, passed to
\code{grid::rasterGrob()}}

\item{coordinates}{\code{data.frame} with \code{rownames} matching the \code{names} in
\code{scores} and columns \code{x} and \code{y} specifying the plotting position of each
observation}
}
\value{
server function, to be passed to \code{shiny::shinyApp()}
}
\description{
SpatialCPie server
}
\keyword{internal}
