% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.makeUI}
\alias{.makeUI}
\title{SpatialCPie UI}
\usage{
.makeUI(imageButton = FALSE)
}
\arguments{
\item{imageButton}{\code{\link[base]{logical}} specifying if the UI should
include a "show image" radio button.}
}
\value{
SpatialCPie UI, to be passed to \code{\link[shiny]{shinyApp}}.
}
\description{
SpatialCPie UI
}
\keyword{internal}
