% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.maximizeOverlap}
\alias{.maximizeOverlap}
\title{Maximize overlap}
\usage{
.maximizeOverlap(xss)
}
\arguments{
\item{xss}{list of lists of labels}
}
\value{
\code{xss}, relabeled so as to maximize the overlap between labels in
consecutive label lists
}
\description{
Maximize overlap
}
\keyword{internal}
