% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.preprocessData}
\alias{.preprocessData}
\title{Preprocess data}
\usage{
.preprocessData(counts, margin, resolutions, assignmentFunction,
  coordinates = NULL)
}
\arguments{
\item{counts}{count matrix.}

\item{margin}{which margin of the count matrix to cluster.}

\item{resolutions}{vector of resolutions to cluster.}

\item{assignmentFunction}{function to compute cluster assignments.}

\item{coordinates}{optional \code{data.frame} with coordinates for each spot.}
}
\value{
list with the following elements:
\itemize{
\item \code{$assignments}: tidy assignments
\item \code{$scores}: cluster scores for each spot in each resolution
\item \code{$colors}: cluster colors
\item \code{$coordinates}: spot coordinates, either from \code{coordinates} or parsed from
\code{assignments}
}
}
\description{
Preprocesses input data for \code{\link{.makeServer}}.
}
\keyword{internal}
