% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.preprocessData}
\alias{.preprocessData}
\title{Preprocess data}
\usage{
.preprocessData(counts, coordinates, margin, resolutions,
  assignmentFunction)
}
\arguments{
\item{counts}{count matrix.}

\item{coordinates}{spot coordinates.}

\item{assignments}{list of assignment vectors.}
}
\value{
list with the following elements:
\itemize{
\item \code{$assignments}: tidy assignments
\item \code{$distances}: sample-centroid distances
\item \code{$colors}: cluster colors
\item \code{$coordinates}: spot coordinates, either from \code{coordinates} or parsed from
\code{assignments}
}
}
\description{
Preprocesses input data for \code{\link{.makeServer}}.
}
\keyword{internal}
