% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\name{.tidyAssignments}
\alias{.tidyAssignments}
\title{Tidy assignments}
\usage{
.tidyAssignments(assignments)
}
\arguments{
\item{assignments}{list of assignment vectors.}
}
\value{
a \code{\link[base]{data.frame}} containing the \code{assignments}, with
the data relabeled so that the overlap between consecutive assignment
vectors is maximized. Additionally, a "root" resolution is added.
}
\description{
Tidy assignments
}
\keyword{internal}
