% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSpotFile.R
\name{parseSpotFile}
\alias{parseSpotFile}
\title{Parsing of spot detector output}
\usage{
parseSpotFile(file)
}
\arguments{
\item{file}{A spot file that contains the pixel coordinates for each spatial area after alignment to the tissue image via the}
}
\value{
FIXME
ST spot detector tool (https://github.com/SpatialTranscriptomicsResearch/st_spot_detector).
}
\description{
A function that parses the output from the ST spot detector tool to receive pixel coordinates for the spatial areas.
}
\examples{
## Create spot file
data <- matrix(
  c(
    c(7, 18, 7.00, 18.07, 563.2, 947.0),
    c(8, 11, 8.00, 11.04, 612.5, 627.7)
  ),
  byrow = TRUE,
  nrow = 2
)
filename <- tempfile()
write.table(
  data,
  file = filename,
  sep = '\\t',
  quote = FALSE,
  col.names = c("x", "y", "new_x", "new_y", "pixel_x", "pixel_y")
)

## Parse spot file
parseSpotFile(filename)

## Delete spot file
unlink(filename)
}
\keyword{spotdetector}
