% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SpatialCPie.R
\docType{package}
\name{SpatialCPie-package}
\alias{SpatialCPie}
\alias{SpatialCPie-package}
\title{SpatialCPie: Cluster analysis of Spatial Transcriptomics data}
\description{
A shiny gadget that can be used for cluster analysis of Spatial
    Transcriptomics data.
}
\author{
\strong{Maintainer}: Joseph Bergenstraahle \email{j.bergenstrahle@scilifelab.se}

}
